+++ core-bundle/tests/Framework/ContaoFrameworkTest.php
@@ -577,4 +577,126 @@ class ContaoFrameworkTest extends TestCase
 
         $this->assertSame($class, $prop->getValue($adapter));
     }
+
+    public function testRegistersTheHookServices(): void
+    {
+        $request = new Request();
+        $request->attributes->set('_route', 'dummy');
+        $request->attributes->set('_scope', ContaoCoreBundle::SCOPE_BACKEND);
+        $request->attributes->set('_contao_referer_id', 'foobar');
+        $request->setLocale('de');
+
+        $container = $this->mockContainerWithContaoScopes();
+        $container->get('request_stack')->push($request);
+        $container->set('test.listener', new \stdClass());
+        $container->set('test.listener2', new \stdClass());
+
+        $GLOBALS['TL_HOOKS'] = [
+            'getPageLayout' => [
+                ['test.listener.c', 'onGetPageLayout'],
+            ],
+            'generatePage' => [
+                ['test.listener.c', 'onGeneratePage'],
+            ],
+            'parseTemplate' => [
+                ['test.listener.c', 'onParseTemplate'],
+            ],
+            'isVisibleElement' => [
+                ['test.listener.c', 'onIsVisibleElement'],
+            ],
+        ];
+
+        $listeners = [
+            'getPageLayout' => [
+                10 => [
+                    ['test.listener.a', 'onGetPageLayout'],
+                ],
+                0 => [
+                    ['test.listener.b', 'onGetPageLayout'],
+                ],
+            ],
+            'generatePage' => [
+                0 => [
+                    ['test.listener.b', 'onGeneratePage'],
+                ],
+                -10 => [
+                    ['test.listener.a', 'onGeneratePage'],
+                ],
+            ],
+            'parseTemplate' => [
+                10 => [
+                    ['test.listener.a', 'onParseTemplate'],
+                ],
+            ],
+            'isVisibleElement' => [
+                -10 => [
+                    ['test.listener.a', 'onIsVisibleElement'],
+                ],
+            ],
+        ];
+
+        $framework = $this->mockContaoFramework(
+            $container->get('request_stack'),
+            $this->mockRouter('/index.html'),
+            [],
+            [],
+            $listeners
+        );
+
+        $framework->setContainer($container);
+
+        if ($framework->isInitialized()) {
+            $reflection = new \ReflectionObject($framework);
+            $reflectionMethod = $reflection->getMethod('registerHooks');
+            $reflectionMethod->setAccessible(true);
+            $reflectionMethod->invoke($framework);
+        } else {
+            $framework->initialize();
+        }
+
+        $this->assertArrayHasKey('TL_HOOKS', $GLOBALS);
+        $this->assertArrayHasKey('getPageLayout', $GLOBALS['TL_HOOKS']);
+        $this->assertArrayHasKey('generatePage', $GLOBALS['TL_HOOKS']);
+        $this->assertArrayHasKey('parseTemplate', $GLOBALS['TL_HOOKS']);
+        $this->assertArrayHasKey('isVisibleElement', $GLOBALS['TL_HOOKS']);
+
+        // Test hooks with high priority are added before low and legacy hooks
+        // Test legacy hooks are added before hooks with priority 0
+        $this->assertSame(
+            [
+                ['test.listener.a', 'onGetPageLayout'],
+                ['test.listener.c', 'onGetPageLayout'],
+                ['test.listener.b', 'onGetPageLayout'],
+            ],
+            $GLOBALS['TL_HOOKS']['getPageLayout']
+        );
+
+        // Test hooks with negative priority are added at the end
+        $this->assertSame(
+            [
+                ['test.listener.c', 'onGeneratePage'],
+                ['test.listener.b', 'onGeneratePage'],
+                ['test.listener.a', 'onGeneratePage'],
+            ],
+            $GLOBALS['TL_HOOKS']['generatePage']
+        );
+
+        // Test legacy hooks are kept when adding only hook listeners with high priority.
+        $this->assertSame(
+            [
+                ['test.listener.a', 'onParseTemplate'],
+                ['test.listener.c', 'onParseTemplate'],
+            ],
+            $GLOBALS['TL_HOOKS']['parseTemplate']
+        );
+
+        // Test legacy hooks are kept when adding only hook listeners with low priority.
+        $this->assertSame(
+            [
+                ['test.listener.c', 'onIsVisibleElement'],
+                ['test.listener.a', 'onIsVisibleElement'],
+            ],
+            $GLOBALS['TL_HOOKS']['isVisibleElement']
+        );
+    }
 }